EC-79 refactor: moving components from user page
