+++ src/app/user/components/customer-info/customer-info.component.spec.ts
