+++ src/app/user/components/dialogs/password-dialog/password-dialog.component.scss
