+++ src/app/user/pages/profile/profile.component.ts
@@ -13,7 +13,7 @@ import { Customer } from '@app/auth/models/customer.model';
 import { selectCustomer } from '@app/ngrx/selectors/customer.selector';
 import { NotificationService } from '@app/shared/services/notofication.service';
 
-import { CustomerActionBuilder } from './services/customer-action-builder.service';
+import { CustomerActionBuilder } from '../../services/customer-action-builder.service';
 
 @Component({
   selector: 'ec-profile',
